Refactored Libraries into seperate headers and src
